sphere.c (gsl_ran_dir_nd): number of dimensions is now unsigned (size_t)
logarithmic.c (gsl_ran_logarithmic_pdf): removed warning about passing
arg 2 of `pow' as floating rather than integer due to prototype

+++ randist/logarithmic.c
@@ -51,7 +51,7 @@ gsl_ran_logarithmic_pdf (const unsigned int k, const double p)
     }
   else 
     {
-      double P = pow(p, k) / (double) k / log(1/(1-p)) ;
+      double P = pow(p, (double)k) / (double) k / log(1/(1-p)) ;
       return P;
     }
 }
